simplify main
